fix: nextjs example export
