+++ examples/framesjs-starter/app/examples/new-api-multi-protocol/frames.ts
@@ -1,0 +1,34 @@
+/* eslint-disable react/jsx-key */
+import { farcasterHubContext, openframes } from "frames.js/middleware";
+import { createFrames } from "frames.js/next";
+import { getXmtpFrameMessage, isXmtpFrameActionPayload } from "frames.js/xmtp";
+import { DEFAULT_DEBUGGER_HUB_URL } from "../../debug";
+
+export const frames = createFrames({
+  basePath: "/examples/new-api-multi-protocol",
+  initialState: {
+    pageIndex: 0,
+  },
+  middleware: [
+    farcasterHubContext({
+      hubHttpUrl: DEFAULT_DEBUGGER_HUB_URL,
+    }),
+    openframes({
+      clientProtocol: {
+        id: "xmtp",
+        version: "2024-02-09",
+      },
+      handler: {
+        isValidPayload: (body: JSON) => isXmtpFrameActionPayload(body),
+        getFrameMessage: async (body: JSON) => {
+          if (!isXmtpFrameActionPayload(body)) {
+            return undefined;
+          }
+          const result = await getXmtpFrameMessage(body);
+
+          return { ...result };
+        },
+      },
+    }),
+  ],
+});
